Add timestamps to messages, cleanup
