+++ src/views/notifications/components/newMessageNotification.js
@@ -12,17 +12,12 @@ import {
   CardContent,
 } from '../../../components/threadFeedCard/style';
 import Icon from '../../../components/icons';
-import { MessageGroupContainer } from '../../../components/messageGroup/style';
-import { AuthorAvatar, AuthorByline } from '../../../components/messageGroup';
-import Message from '../../../components/message';
 import { sortAndGroupNotificationMessages } from './sortAndGroupNotificationMessages';
 import {
   NotificationCard,
   TextContent,
   NotificationListRow,
-  AttachmentsWash,
   SuccessContext,
-  HzRule,
   Content,
 } from '../style';
 
@@ -67,44 +62,6 @@ export const NewMessageNotification = ({
             {' '}
             {actors.asString} {event} {context.asString} {date}{' '}
           </TextContent>
-          <AttachmentsWash>
-            <HzRule>
-              <hr />
-              <Icon glyph="message" />
-              <hr />
-            </HzRule>
-            {messages.map((group, i) => {
-              const initialMessage = group[0];
-              let author = actors.asObjects.filter(
-                user => user.id === initialMessage.senderId
-              )[0];
-              const me = currentUser ? author.id === currentUser.id : false;
-
-              return (
-                <MessageGroupContainer key={i}>
-                  {!me && <AuthorAvatar user={author} />}
-
-                  <MessageGroupContainer me={me}>
-                    <AuthorByline user={author} me={me} />
-                    {group.map((message, i) => {
-                      return (
-                        <Message
-                          key={i}
-                          message={message}
-                          link={`#${message.id}`}
-                          me={me}
-                          canModerate={me}
-                          pending={message.id < 0}
-                          currentUser={currentUser}
-                          context={'notification'}
-                        />
-                      );
-                    })}
-                  </MessageGroupContainer>
-                </MessageGroupContainer>
-              );
-            })}
-          </AttachmentsWash>
         </Content>
       </CardContent>
     </NotificationCard>
